Fixed function override

+++ include/OsmAndCore/Map/IAtlasMapRenderer.h
@@ -28,7 +28,7 @@ namespace OsmAnd
         virtual QVector<TileId> getVisibleTiles() const = 0;
         virtual unsigned int getVisibleTilesCount() const = 0;
         virtual PointD getCameraCoordinates() const = 0;
-        virtual float getCameraHeight() const = 0;
+        virtual double getCameraHeight() const = 0;
 
         virtual float getTileSizeOnScreenInPixels() const = 0;
